Define version for share_access_rules
This is a similar issue as I249b293b11a0e57ddf685de284681f060a334c1c
but with share_access_rules.

Not specifying a microversion when using the Python command line
interpreter with the SDK results in it not functioning properly. The
SDK raises an error stating that the URL cannot be found. However,
when a microversion is explicitly defined, the SDK functions correctly
without any additional configuration.

Change-Id: I6bbad8a74dba781115487882e7a72adf7f394f2f